Create Time Complexity: Primality(Java)
A prime is a natural number greater than 1  that has no positive divisors other than 1  and itself. Given p  integers, determine the primality of each integer and print whether it is Prime or Not prime on a new line.

Note: If possible, try to come up with an O(√n) primality algorithm, or see what sort of optimizations you can come up with for an  O(n)  algorithm. Be sure to check out the Editorial after submitting your code!

Function Description

Complete the primality function in the editor below. It should return Prime if n is prime, or Not prime.

primality has the following parameter(s):

n: an integer to test for primality
Input Format

The first line contains an integer p, , denoting the number of integers to check for primality.
Each of the p subsequent lines contains an integer,n , the number you must test for primality.