Create OrderMangementSystem(Oracle)
A retail company maintain the data of its customers in the CUSTOMER table. 
Write a query to pring the ID's and tthe NAME'S of the customers ,sotred by CUSTOMER.NAME in descending order.
if two or more customers have the same CUSTOMER.NAME, then sort these by CUSTOMER.ID in ascending order.